### 2021-01-04 #2933 @whedon
[PRE REVIEW]: SAR tools: A QGIS plugin for generating SAR descriptors
**Submitting author:** @Narayana-Rao (<a href="http://orcid.org/0000-0002-6496-7283">NR Bhogapurapu</a>)
**Repository:** <a href="https://github.com/Narayana-Rao/SAR-tools" target ="_blank">https://github.com/Narayana-Rao/SAR-tools</a>
**Version:** v0.6
