### 2021-01-04 #2933 @arfon
@Narayana-Rao - please note you have a typo in your paper - affiliations is mis-spelt for the following authors:

```
  - name: Subhadip Dey
    orcid: 0000-0002-4979-0192
    affliation: 1
  - name: Dipankar Mandal
    orcid: 0000-0001-8407-7125
    affliation: 1
  - name: Avik Bhattacharya
    orcid: 0000-0001-6720-6108
    affliation: 1
  - name: Y. S. Rao
    affliation: 1
```

Your paper is also missing the required 'Statement of need' section.

Also, if you want to fix the DOIs (noting that Whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
