### 2021-01-04 #2933 @arfon
:wave: @usethedata happy new year! Would you be willing to edit this submission for JOSS?
