### 2021-01-11 #2933 @hugoledoux
> 👋 @usethedata happy new year! Would you be willing to edit this submission for JOSS?

if @usethedata cannot, I could do this. I teach with QGIS and I know some people in the SAR community
