### 2021-02-04 #2970 @HenrikJanPersson
The plugin is clean and has a well defined purpose. The paper is short but well written, but it is missing several required parts. Additionally, please see my comments below.

1. The purpose is to provide an easy procedure to compute a number of polarimetric indices (parameters) that may be of interest for the user. However, it requires the input data in a format that SNAP and PolSAR Pro delivers, and this SW combination already provides the functionality of deriving the indices that this plugin computes. Therefore, I am still wondering about the need for this plugin?
2. An example of how the PI is used is still missing. Suggest to also tell the user about the sample data.
3. I suggest changing the name from SAR Tools to PolSAR Tools, since they are related to polarimetric SAR, not SAR generally.

Some minor suggestions for improving the PI are listed here:
1. change the name of "View data" to "Select data" or "Open file".
2. provide an information dialog about what happened when the files were generated (where they are located, and possibly also how to open them).
3. add option for selecting the output folder.
4. process bar not updating when computing an index.
5. why is it necessary to both select a folder and a file, independent of each other? Insufficient error message to user if they are not matching.
6. It is not possible to select the "folder" (for whatever use this is?), without first selecting an index. Would be helpful to provide this as information (in the gray folder path area) until an index is selected.
7. cannot change to T-matrix from the pre-selected C-matrix.
8. no  information about what matrix that is expected (e.g., T3/C3) if wrong file is selected
