### 2021-02-08 #2970 @hugoledoux
> The paper is short but well written, but it is missing several required parts.

Just to clarify: the paper is supposed to be short in JOSS, but indeed the points 1 and 2 you bring should be in the paper. As for the new name, this could mean that repository name is changed and other side consequences, so I'll leave it to the authors PolSAR does sound better to me though (but I am not an expert in this field).
