### 2021-02-15 #2970 @Narayana-Rao
Dear @hugoledoux and @HenrikJanPersson,

Thank you very much for your constructive comments. Please find the attached document for a detailed response to your comments.

[Response_to_review_comments_2970.pdf](https://github.com/openjournals/joss-reviews/files/5980352/Response_to_review_comments_2970.pdf)
