### 2021-02-15 #2970 @liberostelios
The paper is well written and quite concise. As a non-expert in the SAR processing field I can't state much about this aspect, although as a GIS user and developer I wonder to which extend there is an overlap between this tool and GRASS (whose processing toolkit can be used from inside QGIS). Do you think it would be useful to provide a simple statement about how GRASS can deal with SAR and how your plugin might relate to this (a sentence or two might suffice)?

The plugin is easy to install and seems to be working for the most part. However, here are some comments about it:
- I ran the plugin against the sample data on Linux and I came across an exception: `AttributeError: module 'os' has no attribute 'startfile'`. Please, beware that the `os` Python package might have differences between operating systems and since QGIS is expected to work well on Windows, macOS and Linux you might want to tackle this differently (see [here](https://stackoverflow.com/questions/29823028/attributeerror-module-object-has-no-attribute-startfile)).
- I find the user interface a bit overwhelming. There is too much information and functionality all in one form. Is this something necessary? For instance, can't the different data types be individual options, instead of just tabs in the same window?
- The QGIS community has been strongly suggesting to move away from plugins with their own UI and towards extending the processing toolkit instead (see a nice [talk](https://www.youtube.com/watch?v=pdO49ysuqVY) about this). In addition to my previous point, I noticed that your UI has many similarities to the interface that custom processing algorithms provide and I think it would require minimal effort to convert your current code into processing algorithms. I think that this would take a lot of workload away from you (you won't have to design and maintain your own GUI) and it would make it easier for users to use the plugin both through the interface and in an automated fashion. There, I suggest that you consider converting this plugin to a processing algorithm instead (at least in the future).
