### 2021-02-25 #2970 @HenrikJanPersson
Thank you for the updates and for constructive feedback on my comments. Several things have been clarified and I think the changes have improved the overall impression greatly. I still have a few important comments.

SAR tools or any other more descriptive name:
I agree with the authors that PolSAR often covers quad-pol data. Yet, I disagree with the authors that PolSAR would be limited to this type. In my experience, Polarimetric SAR intends all types of SAR where the polarization matters and more than a single combination is considered (e.g., HH, VV, or HV). In particular, PolSAR covers the target decomposition the authors are well aware of and also describe in the paper and address with the plugin, therefore it would be a much better fit.
The name "SAR tools" would let the user assume to also get access to basic SAR operations, e.g., multilooking and basic filtering operations, which currently is not the case. I think SAR tools is simply too generic for this very particular set of operations you provide. Assuming you stick with SARtools, I think the background in the paper must ensure to not only referring to free products powered by ESA. Be ware of other softwares too (not limited to PolSAR but SAR in general), like DORIS, ISCE, GMTSAR, ROI_PAC and more.

Paper:
I would suggest to change "... all the three available ..." to only "... the three ...", since there are non-linear polarimetric acquisition modes as well (e.g. circular), which you do not cover.

Detail:
> process bar not updating when computing an index.
>>The progress will get updated for the currently running process (Eg. Computing RVI from Full-pol data),
however, if the dataset is small (like the sample data set), you may not notice the update because the
whole progress bar will be lled and emptied in less than a second.

If it computes this so fast that I cannot notice, why not leaving it at 100% when finished? I see no point of resetting it to 0% before a new Process has been activated.
