### 2021-03-08 #2970 @hugoledoux
@Narayana-Rao is it clear to you that the review is iterative? I am asking because you wrote a formal response to reviewer in a PDF, and while it's fine, I believe it would be better if you just answered here in plain text and discussed with the reviewers.

For the name, as I said, this is up to you, but as @HenrikJanPersson  pointed out, it would be a better name and furthermore I reckon that since you're still in beta it's still time to change.

The comments of @liberostelios about the fact that the plugin is not cross-platform should definitively be addressed, and as for the GUI, I would also encourage you to use the "standard" QGIS one (it will also make is easier to upgrade the code in the future!).

Let us now which timetime you have in mind to address those issues.
