### 2021-03-15 #2970 @hugoledoux
@Narayana-Rao Great that you changed the named 👍

For the GUI, I agree with @liberostelios that a processing interface would be better, but also realise that perhaps it is asking too much. However, I think you have to at least harmonise with other plugins, and provide an "OK/Run" button at the bottom right, and a "Cancel" button too. Right now there is no cancel button whatsoever. Also, the help should be a "Help" button at the bottom-left, like for virtually all plugins.

See for instance this one:
![IMAGE 2021-03-15 15:22:46](https://user-images.githubusercontent.com/1546518/111168362-4c865400-85a2-11eb-8667-e17a897dbab3.jpg)

Your "process" button is not a button, and will create confusion in my opinion, see when I enlarge the window under macOS:
![CleanShot 2021-03-15 at 15 22 00@2x](https://user-images.githubusercontent.com/1546518/111168340-455f4600-85a2-11eb-8c21-6489a2c4d757.png)

---

@liberostelios & @HenrikJanPersson : some of the checkboxes at the top are not clicked yet. Is it because the author needs to add something? Or is it because you haven't had time yet to look at this?
