### 2021-03-24 #2970 @Narayana-Rao
Dear @hugoledoux, I want to thank you for positively responding to our "response to review comments" on behalf of our team. I appreciate your suggestions regarding the UI. I agree that our tool/plugin should have standard UI elements, which will enhance usability and avoid confusion that may arise due to a non-standard UI. Therefore, according to your suggestions, we have now updated the plugin UI in the recent release [(v0.7)](https://github.com/Narayana-Rao/PolSAR-tools/releases/tag/v0.7).
