### 2021-03-24 #2970 @Narayana-Rao
> I agree with @hugoledoux comments about GUI, but other than that I can raise any big objections on the publication of the paper. So I've filled all checkboxes from my side.

Dear @liberostelios, Thank you very much for accepting the paper on the PolSAR tools plugin.  Your comments on the plugin helped us a lot in improving the overall quality of the plugin.
