### 2021-04-08 #2970 @hugoledoux
OK, we're getting to the end of this review.

I went over the paper and fixed a few things, could you check it and merge (if you agree):
https://github.com/Narayana-Rao/PolSAR-tools/pull/7
