### 2021-04-08 #2970 @Narayana-Rao
> OK, we're getting to the end of this review.
>
> I went over the paper and fixed a few things, could you check it and merge (if you agree):
> [Narayana-Rao/PolSAR-tools#7](https://github.com/Narayana-Rao/PolSAR-tools/pull/7)

Dear @hugoledoux, Thank you very much for your edits. I have checked and merged your pull request [Narayana-Rao/PolSAR-tools#7](https://github.com/Narayana-Rao/PolSAR-tools/pull/7).
