### 2021-04-08 #2970 @Narayana-Rao
> OK, both reviewers recommend acceptance so we're moving towards this.
>
> At this point could you:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission.

Dear @hugoledoux,

I have finished the archival of the repository according to the given guidelines. Please verify and confirm the following check list.

1. Version [v0.7](https://github.com/Narayana-Rao/PolSAR-tools/releases/tag/v0.7)
2. Zenodo archive [10.5281/zenodo.4621292](https://doi.org/10.5281/zenodo.4621292)
3. I have verified the archive's metadata and also added ORCIDs authors.
4. Zenodo archive [10.5281/zenodo.4621292](https://doi.org/10.5281/zenodo.4621292)
