### 2021-04-08 #2970 @Narayana-Rao
> > Dear @danielskatz, Could you please verify the rendered .pdf file. For me it's showing 404 not found if I click on the DOI link.
>
> It works fine for me now. Sometimes if you try it too soon, it doesn't work, and that gets cached in your system/network. I'm going to close this issue now, since it does work for me, but if it still isn't working for you in 24 hours, then reopen it.

Dear @danielskatz, Thank you very much for your time, edits and prompt responses.
